changing minor inaccuracies

+++ src/Engine/Camera/BasicCameraController/BasicCameraController.h
@@ -5,19 +5,19 @@
 // TODO Movable
 class BasicCameraController : public CameraController {
 private:
-    double movementSpeed;
-	double sensitivity;
+    float movementSpeed;
+	float sensitivity;
 
-    void updateMouseInput(const glm::dvec2& offset);
+    void updateMouseInput(const glm::vec2& offset);
 
 public:
-    BasicCameraController(Camera& camera, double movementSpeed, double sensitivity = 0.07)
+    BasicCameraController(Camera& camera, float movementSpeed, float sensitivity = 0.07f)
             : CameraController(camera), movementSpeed(movementSpeed), sensitivity(sensitivity) {}
     
     ~BasicCameraController() override = default;
 
-	void move_in_direction(const Direction direction, double dt) override;
+	void move_in_direction(const Direction direction, float dt) override;
 
-	void update(const glm::dvec2& offset, double dt) override;
+	void update(const glm::vec2& offset, float dt) override;
 };
