+++ src/Engine/Camera/Dron/Dron.h
@@ -4,21 +4,21 @@
 
 class Dron : public CameraController {
 private:
-    double movementSpeed;
-	double sensitivity;
+    float movementSpeed;
+	float sensitivity;
 
-    glm::dvec2 prev_offset;
+    glm::vec2 prev_offset;
 
-    void updateMouseInput(const glm::dvec2& offset);
+    void updateMouseInput(const glm::vec2& offset);
 
 public:
-    Dron(Camera& camera, double movementSpeed, double sensitivity = 0.07)
+    Dron(Camera& camera, float movementSpeed, float sensitivity = 0.07)
             : CameraController(camera), movementSpeed(movementSpeed), sensitivity(sensitivity) {}
     
     ~Dron() override = default;
 
-	void move_in_direction(const Direction direction, double dt) override;
+	void move_in_direction(const Direction direction, float dt) override;
 
-	void update(const glm::dvec2& offset, double dt) override;
+	void update(const glm::vec2& offset, float dt) override;
 };
